requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
langchain==0.3.15
langchain-openai==0.0.8
openai==1.58.1
langchain-anthropic==0.3.4
langsmith==0.3.1
pydantic-settings==2.7.1
pydantic==2.7.4
pydantic-core==2.18.4
langchain-community==0.3.15
python-dotenv==1.0.1
